rename driver_registry_helper.h
