+++ netebpfext/net_ebpf_ext_bind.c
@@ -8,7 +8,7 @@
 
 #define INITGUID
 
-#include "driver_registry_helper.h"
+#include "ebpf_store_helper.h"
 #include "net_ebpf_ext_bind.h"
 
 //
