+++ netebpfext/net_ebpf_ext_xdp.c
@@ -8,7 +8,7 @@
 
 #define INITGUID
 
-#include "driver_registry_helper.h"
+#include "ebpf_store_helper.h"
 #include "net_ebpf_ext_xdp.h"
 
 //
